Слияние с ревизией(ями) 1130 из OC: added users online feature.
git-svn-id: http://subversion.assembla.com/svn/ocstoreru@521 44b37dae-b762-45cd-8c5e-82695f3631dd

+++ trunk/catalog/controller/product/category.php
@@ -48,9 +48,9 @@ class ControllerProductCategory extends Controller {
 		
 			foreach ($parts as $path_id) {
 				if (!$path) {
-					$path = $path_id;
+					$path = (int)$path_id;
 				} else {
-					$path .= '_' . $path_id;
+					$path .= '_' . (int)$path_id;
 				}
 									
 				$category_info = $this->model_catalog_category->getCategory($path_id);
@@ -64,7 +64,7 @@ class ControllerProductCategory extends Controller {
 				}
 			}		
 		
-			$category_id = array_pop($parts);
+			$category_id = (int)array_pop($parts);
 		} else {
 			$category_id = 0;
 		}
